requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
python-dotenv
google-genai
gTTS
httpx
